memory_palace/requirements.txt
anthropic~=0.45.2
aiofiles==23.2.1
fastapi==0.110.0
Jinja2==3.1.3
openai==1.12.0
python-dotenv==1.0.1
python-multipart==0.0.9
uvicorn==0.27.1
websockets==12.0
boto3~=1.36.11
sentence-transformers
uuid~=1.30
numpy~=2.0.2
gunicorn==20.1.0
botocore~=1.36.12
redis~=5.2.1
awscli